Modifications
I added code on the Employee.js and modified it on the rest of the lib to represent each of their js and their functions to make the code work.  As for the template.html, that was just to test my code and how it would look is not connected to the main code that is just my playground.