Final Details
All of the challenge is all done everything works like it suppose to do and including on the readme, I will add the link of the video with the screenshot of the output of the code and how it looks on the html.  I also added the license I use to make this code happen.